Allow `options` to be passed into `--missing`
This implementation allows people to use `dependency-check . --no-dev`, which will error if a dependency is missing in production. There's one semantic change, which appears to be an existing undiscovered bug/feature, in that `--extra` was not previously checking `peerDependencies`. I've made that change as part of this PR as I believe `peerDependencies` should be treated as dependencies for the semantics of checking usages, but can revert if needed.